tests/wpt/web-platform-tests/html/webappapis/system-state-and-capabilities/the-navigator-object/content/003.xhtml

<?xml version="1.0" encoding="utf-8"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>Content type case insensitivity in isContentHandlerRegistered</title>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script type="application/ecmascript">
test(function()	{
 var ctype = 'application/x-notRegisteredInOtherTCs-003', url = location.href.replace(/\/[^\/]*$/, "") + '/%s';
 navigator.registerContentHandler(ctype, url, 'test');
 assert_equals(navigator.isContentHandlerRegistered(ctype.toUpperCase(), url), 'declined');
});
</script>
</head>
<body>
<div id="log"></div>
</body>
</html>
